validator: make common a module and move the Alternator harness into it
The Alternator harness sat in the module that also declares the
alternator group, so anything outside that group's subtree could only
reach it as a child.

Turn common into a directory module and add the harness as
common::alternator, leaving the alternator module with its group alone.
The items the group modules use are pub(crate) now that they are named
from outside the harness.
