+++ crates/validator/src/alternator/put_item.rs
@@ -4,10 +4,10 @@
  */
 
 use crate::TestActors;
-use crate::alternator;
-use crate::alternator::Item;
-use crate::alternator::TableContext;
 use crate::common;
+use crate::common::alternator;
+use crate::common::alternator::Item;
+use crate::common::alternator::TableContext;
 use aws_sdk_dynamodb::types::AttributeValue;
 use std::sync::Arc;
 use tracing::info;
@@ -152,7 +152,7 @@ async fn put_item_with_invalid_vector_is_not_indexed(actors: Arc<TestActors>) {
 e2etest::group!(
     name = put_item,
     fixtures = (Fixture),
-    parent = alternator::alternator
+    parent = super::alternator
 );
 
 struct Fixture {
