+++ crates/validator/src/alternator/query.rs
@@ -4,12 +4,12 @@
  */
 
 use crate::TestActors;
-use crate::alternator;
-use crate::alternator::Item;
-use crate::alternator::JsonBodyInjectInterceptor;
-use crate::alternator::TableContext;
-use crate::alternator::TableShape;
 use crate::common;
+use crate::common::alternator;
+use crate::common::alternator::Item;
+use crate::common::alternator::JsonBodyInjectInterceptor;
+use crate::common::alternator::TableContext;
+use crate::common::alternator::TableShape;
 use aws_sdk_dynamodb::client::customize::CustomizableOperation;
 use aws_sdk_dynamodb::operation::query::QueryError;
 use aws_sdk_dynamodb::operation::query::QueryOutput;
@@ -752,7 +752,7 @@ async fn query_with_filter_expression(actors: Arc<TestActors>) {
 e2etest::group!(
     name = query,
     fixtures = (Fixture),
-    parent = alternator::alternator
+    parent = super::alternator
 );
 
 struct Fixture {
