+++ crates/validator/src/alternator/ttl.rs
@@ -10,12 +10,12 @@
 //! from the index.
 
 use crate::TestActors;
-use crate::alternator;
-use crate::alternator::Item;
-use crate::alternator::TableContext;
-use crate::alternator::TableShape;
 use crate::alternator::query::QueryBuilderExt;
 use crate::common;
+use crate::common::alternator;
+use crate::common::alternator::Item;
+use crate::common::alternator::TableContext;
+use crate::common::alternator::TableShape;
 use aws_sdk_dynamodb::types::AttributeValue;
 use aws_sdk_dynamodb::types::ScalarAttributeType;
 use aws_sdk_dynamodb::types::Select;
@@ -189,11 +189,7 @@ async fn ttl_expiration_verified_via_query_with_all_projected(actors: Arc<TestAc
     info!("finished");
 }
 
-e2etest::group!(
-    name = ttl,
-    fixtures = (Fixture),
-    parent = alternator::alternator
-);
+e2etest::group!(name = ttl, fixtures = (Fixture), parent = super::alternator);
 
 struct Fixture {
     actors: Arc<TestActors>,
